+++ src/config.ts
@@ -8,27 +8,50 @@ export enum ConfigKeys {
     deckType = 'deckType',
     nextDraw = 'nextDraw',
     enableNotification = 'enableNotification',
-    pileUp = 'pileUp'
+    pileUp = 'pileUp',
+    difficultyLevel = 'difficultyLevel',
+    aggregatePile = 'aggregatePile'
 }
 
 export class Config {
+    static getConfigValue(key: ConfigKeys): unknown {
+        return vscode.workspace.getConfiguration(CONFIG_NAME).get(key);
+    }
+
     static get pickEvery(): number {
-        return Number(vscode.workspace.getConfiguration(CONFIG_NAME).get(ConfigKeys.pickEvery));
+        return Number(Config.getConfigValue(ConfigKeys.pickEvery));
     }
 
     static get isBadgeNotificationEnabled(): boolean {
-        return Boolean(vscode.workspace.getConfiguration(CONFIG_NAME).get(ConfigKeys.enableBadgeNotification));
+        return Boolean(Config.getConfigValue(ConfigKeys.enableBadgeNotification));
     }
     
     static get isNotificationEnabled(): boolean {
-        return Boolean(vscode.workspace.getConfiguration(CONFIG_NAME).get(ConfigKeys.enableNotification));
+        return Boolean(Config.getConfigValue(ConfigKeys.enableNotification));
     }
 
     static get deckType(): string {
-        return String(vscode.workspace.getConfiguration(CONFIG_NAME).get(ConfigKeys.deckType));
+        return String(Config.getConfigValue(ConfigKeys.deckType));
     }
 
     static get pileUp(): boolean {
-        return Boolean(vscode.workspace.getConfiguration(CONFIG_NAME).get(ConfigKeys.pileUp));
+        return Boolean(Config.getConfigValue(ConfigKeys.pileUp));
+    }
+
+    static get aggregatePile(): boolean {
+        return Boolean(Config.getConfigValue(ConfigKeys.aggregatePile));
+    }
+
+    static get difficultyLevel(): number {
+        const difficultyLevel = Config.getConfigValue(ConfigKeys.difficultyLevel) as number;
+
+        if(difficultyLevel < 1) {
+            return 1;
+        }
+        if(difficultyLevel > 5) {
+            return 5;
+        }
+
+        return difficultyLevel;
     }
 }